fix: fix structure and compress and decompress
