+++ src/printWorkingDirChange.js
@@ -1,0 +1,5 @@
+const printCurrentWorkingDirectory = () => {
+  console.log(`You are currently in ${process.cwd()}`);
+};
+
+export default printCurrentWorkingDirectory;
